Freelancer Project Management Platform/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
pydantic-settings==2.6.1
email-validator==2.2.0
